+++ com.unity.postprocessing/PostProcessing/Runtime/Effects/AmbientOcclusion.cs
@@ -127,6 +127,12 @@ namespace UnityEngine.Rendering.PostProcessing
         [Range(1f, 10f), Tooltip("This modifies the thickness of occluders. It increases the size of dark areas and also introduces a dark halo around objects.")]
         public FloatParameter thicknessModifier = new FloatParameter { value = 1f };
 
+        /// <summary>
+        /// Add a bias distance to sampled depth in AO to reduce self-shadowing aliasing artifacts.
+        /// </summary>
+        [Range(0f, 0.001f), Tooltip("Add a bias distance to sampled depth in AO to reduce self-shadowing aliasing artifacts. ")]
+        public FloatParameter zBias = new FloatParameter { value = 0.0001f };
+
         // HDRP-only parameters
 
         /// <summary>